refactor: 클래스 명 변경, 팩토리 위치 이동 - CustomResponseCode -> ErrorResponseCode - ErrorResponseCode 위치를 exception 팩토리로 이동
